cv_ssl/requirements.txt
torch>=1.10.0
torchvision>=0.11.1
einops
lightning==2.1.2
torchmetrics>=0.6.0, <0.12.0
tqdm
wandb
scipy
timm
scikit-learn
hydra-core
h5py